+++ src/msg/warn.cc
@@ -112,8 +112,8 @@ void Warn::nondeterministic_tags(
         } else {
             fprintf(stderr, "tag '%s'", tagname);
         }
-        fprintf(stderr, " %shas %u%s degree of nondeterminism", incond(cond).c_str(),
-                static_cast<uint32_t>(nver), nver == 2 ? "nd" : nver == 3 ? "rd" : "th");
+        fprintf(stderr, " %shas %zu%s degree of nondeterminism",
+                incond(cond).c_str(), nver, nver == 2 ? "nd" : nver == 3 ? "rd" : "th");
         msg.warning_end(names[NONDETERMINISTIC_TAGS], e);
     }
 }
